requirements.txt
langchain
sentence-transformers
faiss-cpu
numpy
pandas
fastapi
uvicorn
pydantic
python-dotenv
